Nuevos cambios
Creé UsuarioPartidaService y su implementacion.. hice el procesar IntentoconAjax en controllerJuego y esto me llevo a modificar varias clases, servicios, repositorios..

+++ src/main/java/com/tallerwebi/dominio/model/Ronda.java
@@ -16,7 +16,7 @@ public class Ronda {
     @OneToOne
     private Palabra palabra;
     private String definicion;
-    private int numeroDeRonda;
+    private Long numeroDeRonda;
     private Estado estado;
     private LocalDateTime fechaHora; //LocalDateTime es una clase que se mapea a datetime con huibernate, con .now() genera el horario actual.
 
@@ -44,11 +44,11 @@ public class Ronda {
         this.palabra = palabra;
     }
 
-    public int getNumeroDeRonda() {
+    public Long getNumeroDeRonda() {
         return numeroDeRonda;
     }
 
-    public void setNumeroDeRonda(int numeroDeRonda) {
+    public void setNumeroDeRonda(Long numeroDeRonda) {
         this.numeroDeRonda = numeroDeRonda;
     }
